Create even spacing between card border and title
Because of default CSS of browser, title has top and bottom margins of 21.44. As a result, the top and bottom of title box is about 21.44 wide, while the left and right is 10px (from .card padding).

The result I want is 10px wide top, bottom, left & right.
The extra top margin of 21.44px is from browser defaults, so setting margin-top: 0px will solve the uneven space issue for top of title.
The bottom of title is larger than 10px due to 21.44px defaults. Setting margin-bottom: 10px does the trick.