Reformat links and their containers
I removed the bullets, padding, and margin of each link.

I also removed the underlines of each link and changed their color to a dark gray.

What's left is adding some padding to create space between viewport and the links, and repositioning the links so that some are on the left and some are on the right.