Push footer to the bottom of window
I applied display:flex to <body> and declared flex-flow: column, so that the vertical positioning stays the same.

Then, by applying flex: 1 1 auto to "content" container, I allow it to grow into the remaining vertical space. This growth pushes the footer towards the bottom of the window, as it is inflexible.